refactor(biggest): split utils for 450-line rule
